Fix encoder only board issues for manufacturing;
The board was a couple of mils too wide to get 3x3 boards on a panel.
Thus the components were moved a bit to the left and traces squeezed to
fit in a smaller space.

Also there were some duplicate texts in the jumper "legends" below the
encoder and those texts were removed.

The hole on the right top corner was moved to a symmetrical position
compared to the one on the left top.