Make it possible to use buttons also on L2 and R2;
The previous design only allowed for one of the boards in the chain
to have a button (basically the one closest to the connector board).
The latest PCB design with it's voltage divider setup enables the use
of two buttons. However if both buttons are pressed at the same time,
there might not be an interrupt generated on the second press and also
reading the value from the line might not work correctly.