Add wait before I2C initialization to prevent race condition with master;
Before this change the slave would only get the address from the master
if the slave was configured so that it needed to setup all boards as
encoder boards. When trying to initialize the board configured to have
only a button matrix, it seemed to fail to get the address and caused
the whole bus to freeze. The idea in this fix is to give the master some
time to boot before starting to request addresses from it.