Update the FFC adapter to use bottom contact connectors;
The previous design had the pin orders to match those in the connector
board. When using a forward direction cable (contacts on the same side
on both ends) and a top contact connector on both ends, the pin order
would get mirrored. Because the forward direction cable and the top
contact connectors are (supposedly?) the norm, it's probably better to
just reverse the pins on the adapter in order to be able to use the same
connectors and cables everywhere.