Separate application specific code and "libary" code;
In order to be able to use the slave code as a library for different
projects without needing to fork the whole codebase, the common setup,
event handling, etc. logic should be separated from the application
specific logic. This commit tries to do just that, but does not
completely succeed. The generic setup code depends on the application
specific setup code (i.e. config.h) and thus that file has to be
included in both the slave.h and slave.cpp.