Prevent button control number collisions;
If there are multiple buttons on the same board (i.e. pads or a button
matrix) there needs to be something to differentiate the button events
from different buttons. This commit adds board * 20 to the control/input
value in order to enable max 20 buttons on each board (which is way more
than currently possible).