Share single instance of Adafruit_NeoPixel between LED chains;
The MCU ran out of memory when using three Adafruit_NeoPixel's with a
total of 64 LEDs. This commit changes the implementation so that only
one instance of the class is active at one time. This however means that
the sketch has to initialize all the LEDs in the chain when a new
NeoPixelBus is initialized (when changing the state of a different LED
chain).